Changed `Text to carry a string list.
As pointed out by @dbuenzli in #1, long sequences of text will exhaust
the 16MB limit on OCaml string length on 32-bit systems. After this
change, the parsers split long strings in the input into chunks of
legnth approximately Sys.max_string_length / 2. On 32-bit platforms,
that comes out to about 8MB. On 64-bit platforms, that comes out to
about 64TB, which doesn't seem likely to be reached.

It is actually still possible to reach the string length limit in other
places in the parsers, for example using extremely long comments or XML
entity references. That will be addressed in future commits.

Resolves #1.