Deprecate Markup_lwt.ensure_tail_calls
This was a workaround. It was never necessary, but this is difficult to
immediately see: Lwt.on_any t f g does not do tail calls. However, this
matters only when t is already resolved. When t is still pending, it is
not possible to get a tail call of f or g anyway. In that case, stack
growth is controlled by whether one is using Lwt.wakeup_later, or, if
using Lwt.wakeup, by the interleaving of adding waiters to t and calls
to Lwt.wakeup.

Therefore, it is sufficient to check the state of t, tail-call f and g
manually if t is resolved, and use the existing Lwt.on_any if t is
pending.

Closes ocsigen/lwt#206.