Speed: trie for entity lookup
The trie currently uses considerable memory, on the order of 7MB on a
64-bit system. This can be optimized in the future, probably so that it
uses on the order of 1MB. Alternatively, the trie can be replaced by
another data structure, if necessary.

The trie is created lazily, so users not parsing HTML, or parsing HTML
without entity references, do not pay this penalty.