Speed: buffer channel reads
The speedup is insensitive to the buffer size once it is over 64 bytes.
Presumably, that is the point at which the time to read the bytes
dominates the overhead of making the read call and handling its result.
However, I arbitrarily chose 4096-byte buffers.

Time to run this code

  Markup.(file "foo.ml" |> fst |> drain)

decreased by about 30%. Testing was done with Flambda -O3.

Related #15.