Fix dependency on lwt.unix
Module Markup_lwt_unix depends on Findlib packages lwt and lwt.unix.
However, the only dependency actually listed in _tags was on package
lwt. Until recently, Lwt installed all files in a flat directory.
Depending on package lwt causes ocamlfind to add a -I option for that
directory; since the directory was flat, that option was sufficient to
be able to accidentally find the files of lwt.unix as well.

Lwt now installs files in a directory hierarchy. In particular, lwt.unix
is now installed in a separate subdirectory. The -I option implied by
depending on package lwt is therefore no longer accidentally sufficient
for using lwt.unix, and this causes an error during compilation of
Markup.ml if Lwt is present.

So, this commit does the right thing, and adds the proper, explicit
dependency on lwt.unix.

This bug was originally found in Markup.ml, as well as several other
packages, by @andrewray during ocsigen/lwt#374.

Fixes #18.
See also ocsigen/lwt#401.