Handle </html> tag as per the specification
An unfortunate side effect of this change is that the parser now always
ignores </html> tags, because the point where it would legitimately
close an <html> element cannot be distinguished from an error. That
would require adding more state or callback parameters, which was not
done in this commmit.

As a result, <html> elements are only closed implicitly by the end of
input. As a further result, the location of the closing </html> tag is
always reported as the location of the end of the input. In other words,
if there is a legitimate </html> tag, its location is not accurately
reported.

Fixes #28.