Handle </body> as per the specification
This is similar to the preceding commit, and has the same side effect
that essentially the only common way to close a <body> element is by
ending input. The </body> tag is ignored for closing purposes. As a
result, element-ending </body> signals are reported to be at the end of
input, instead of where the </body> tag appears in the input, even if
the </body> tag is in a syntactically-valid position.

Part of #28.