HTML writer: preserve the doctype
The HTML5 spec requires outputting <!DOCTYPE html> as the doctype.
However, Markup.ml already has a helper (Markup.html5) for replacing any
doctype with the HTML5 doctype. So, it is better for the writer to
output whatever doctype is present in the signal stream, trusting the
user to call Markup.html5 if it is necessary for their purpose.

Part of aantron/lambdasoup#37.